+++ packages/banana-react/src/index.ts
@@ -15,6 +15,8 @@ import { Progress } from './progress';
 import { Rating } from './rating';
 import { Select } from './select';
 import { SelectOption } from './select-option';
+import { Stepper } from './stepper';
+import { Tooltip } from './tooltip';
 
 export {
   Button,
@@ -34,4 +36,6 @@ export {
   Rating,
   Select,
   SelectOption,
+  Stepper,
+  Tooltip,
 };