+++ packages/banana/src/carousel/index.ts
@@ -89,7 +89,7 @@ export default class BCarousel extends LitElement {
   slidesPerView = 1;
 
   private get _slidesPerView() {
-    return this._slides.length >= this.slidesPerView || !this.fill ? this.slidesPerView : this._slides.length;
+    return this._slides.length >= this.slidesPerView || this.disableFill ? this.slidesPerView : this._slides.length;
   }
 
   @property({ type: Boolean, reflect: true })
@@ -111,9 +111,9 @@ export default class BCarousel extends LitElement {
   @property({ type: Number, reflect: true })
   gap = 0;
 
-  // If the `fill` property is set to true, then carousel will be filled when the slide count is less than the `slidesPerView` property.
-  @property({ type: Boolean, reflect: true })
-  fill = true;
+  // By default, carousel will be filled when the slide count is less than the `slidesPerView` property.
+  @property({ type: Boolean, reflect: true, attribute: 'disable-fill' })
+  disableFill = false;
 
   @property({ type: Boolean, reflect: true })
   autoHeight = false;