+++ packages/banana/src/index.ts
@@ -15,6 +15,8 @@ import BProgress from './progress';
 import BRating from './rating';
 import BSelect from './select';
 import BSelectOption from './select-option';
+import BStepper from './stepper';
+import BTooltip from './tooltip';
 
 export {
   BButton,
@@ -34,4 +36,6 @@ export {
   BRating,
   BSelect,
   BSelectOption,
+  BStepper,
+  BTooltip,
 };