+++ packages/banana/src/overlay/index.ts
@@ -1,5 +1,5 @@
 /* eslint-disable lit-a11y/click-events-have-key-events */
-import { LitElement, html, PropertyValues, CSSResultGroup, PropertyValueMap } from 'lit';
+import { CSSResultGroup, LitElement, PropertyValueMap, PropertyValues, html } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 import styles from './index.styles';
 
@@ -13,6 +13,9 @@ export default class BOverlay extends LitElement {
 
   static styles: CSSResultGroup = styles;
 
+  // A style element to set the overflow of body to hidden.
+  private _styleElement = document.createElement('style');
+
   protected willUpdate(_changedProperties: PropertyValueMap<this>): void {
     if (_changedProperties.has('zIndex')) {
       this.style.zIndex = String(this.zIndex);
@@ -22,7 +25,7 @@ export default class BOverlay extends LitElement {
   protected updated(_changedProperties: PropertyValues<this>): void {
     // _changedProperties is the previous values, so we use this.open to get the current value.
     if (_changedProperties.has('open') && this.open) {
-      document.body.style.overflow = 'hidden';
+      document.head.appendChild(this._styleElement);
       document.addEventListener('touchstart', this._preventTouchEvent);
       window.addEventListener('keydown', this._handleEscape);
     } else {
@@ -33,6 +36,7 @@ export default class BOverlay extends LitElement {
   connectedCallback() {
     super.connectedCallback();
     this.style.zIndex = String(this.zIndex);
+    this._styleElement.innerHTML = 'body { overflow: hidden; }';
   }
 
   disconnectedCallback() {
@@ -41,7 +45,7 @@ export default class BOverlay extends LitElement {
   }
 
   private removeEvents() {
-    document.body.style.overflow = 'auto';
+    this._styleElement.remove();
     document.removeEventListener('touchstart', this._preventTouchEvent);
     window.removeEventListener('keydown', this._handleEscape);
   }