+++ packages/banana-react/CHANGELOG.md
@@ -1,5 +1,67 @@
 # @banana/banana-react
 
+## 1.14.2
+
+### Patch Changes
+
+- Fix a Countdown bug causing an exception when the time is set to 0 in separate mode, resulting in missing separators.
+- Updated dependencies
+  - banana-ui@1.14.2
+
+## 1.14.1
+
+### Patch Changes
+
+- Updated babel config.
+- Updated dependencies
+  - banana-ui@1.14.1
+
+## 1.14.0
+
+### Minor Changes
+
+- add stepper component
+
+### Patch Changes
+
+- Updated dependencies
+  - banana-ui@1.14.0
+
+## 1.13.3
+
+### Patch Changes
+
+- Version updated.
+- Updated dependencies
+  - banana-ui@1.13.3
+
+## 1.13.2
+
+### Patch Changes
+
+- Prevent scrolling when overlay is opened in a better way.
+- Updated dependencies
+  - banana-ui@1.13.2
+
+## 1.13.1
+
+### Patch Changes
+
+- Change the fill parameter of carousel to disableFill.
+- Updated dependencies
+  - banana-ui@1.13.1
+
+## 1.13.0
+
+### Minor Changes
+
+- New component——tooltip
+
+### Patch Changes
+
+- Updated dependencies
+  - banana-ui@1.13.0
+
 ## 1.12.5
 
 ### Patch Changes